Update and Reorganize Contribution Guide; Simplify Resources Data Structure (#20)
* new resources in icons, replace base64 images, and Update URL parameter from 'rel' to 'ref' in href.

* feat: Update home page layout and add featured guides card

- Changed "Featured Resources" to "Explore Top Resources" on home page
- Added a new section for "Learn the Basics" with relevant content
- Included a featured guides card with highlight information
- Updated button link text on home page from "Getting started" to "Learn"

* GitButler Workspace Commit

This is a merge commit the virtual branches in your workspace.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - v2/landing-page (refs/gitbutler/v2/landing-page)
   branch head: d2dc5ac017773adc187a0ae84116721eb72a841e
   - app/(home)/page.tsx
   - components/shared/footer.tsx
   - public/gradient.png
   - components/shared/home/guides-card.tsx
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch

* fix: root container margin in homepage

- disabled global colors of fumadocs

* 1.5.2

* add: html overview and new image section component

* 1.5.3

* feat: Add contribution guide (not yet in sidebar) and reorder resources alphabetically

* new favicon

fix: image section not centered in small screens

* 1.5.4

* Add variables and operators guide

* refactor: Remove duplicate iconPath and unused colors section

* refactor: Update resource categories and remove unused properties

This commit updates the resource categories in the `resourcesData` object in the `resources.ts` file. It removes the `thumbnail` and `subtitle` properties from each category. The changes also include removing the `contribution-guide.mdx` file and updating the `meta.json` file to include the new resource page.

---------

Co-authored-by: GitButler <gitbutler@gitbutler.com>